YotubeTextSummarization/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf

faiss-cpu
youtube-transcript-api

langchain-huggingface
transformers
huggingface-hub

python-dotenv

streamlit

